Add files via upload
The SQL is used to modify the dbo.[Sales-Report-CSV] table. The code first drops the Revenue column from the table. It then splits the Transaction Date column into two columns, Date and Time. The code then adds the two columns, Date and Time, to the table. Finally, the code updates the Date and Time columns with the appropriate date and time values from the original Transaction Date column. The code then selects all of the values from the updated table to check that the modifications were successful. In Power Query, I used duplicate removal for Email addresses as many Email Addresses were more than 1 time.